+++ projects/code_scratch/config_handler.py
@@ -6,18 +6,7 @@ class ConfigHandler(ConfigLifecycleHandler):
     """A handler to customize callbacks and tools for different phases."""
 
     def task_begin(self, config: DictConfig, tag: str) -> DictConfig:
-        if tag == 'Architecture':
-            config.callbacks = ['callbacks/coding_callback']
-        elif tag == 'Refiner':
-            config.callbacks = [
-                'callbacks/eval_callback', 'callbacks/coding_callback'
-            ]
-        elif 'worker' in tag:
+        if 'worker' in tag:
             config.callbacks = ['callbacks/artifact_callback']
             delattr(config.tools, 'split_task')
-            config.tools.file_system = DictConfig({
-                'mcp':
-                False,
-                'exclude': ['create_directory', 'write_file', 'list_files']
-            })
         return config
