--- projects/code_scratch_test2/architecture.yaml
@@ -1,62 +1,0 @@
-llm:
-  service: openai
-  model: claude-3-7-sonnet-20250219
-  openai_api_key:
-  openai_base_url: https://dashscope.aliyuncs.com/compatible-mode/v1
-
-generation_config:
-  top_p: 0.6
-  temperature: 0.2
-  top_k: 20
-  stream: true
-  extra_body:
-    dashscope_extend_params:
-      provider: idealab
-    enable_thinking: false
-
-
-prompt:
-  system: |
-    You are a senior software architect. Your responsibility is to break down complex requirements into implementable modules. You need to follow these instructions:
-
-    1. Analyze the original requirement, output a functionality PRD, then a detailed architectural design:
-      * You may add extra possible and reasonable functionalities to rich the project.
-      * Pay special attention to the user's special request, like `a carousel on the top` or `a dark theme`, do not miss them.
-      * Determine the programming language.
-      * Design all modules, making sure not to miss any functionality.
-      * The http interfaces and detailed data structures(arguments, responses)
-    
-    2. Code language preferences:
-      * Frontend: React + vite.js, do not use ts check in package.json
-      * Backend: Node.js
-      * Algorithm: Python
-    
-    3. Think carefully, list the file structure from backend to frontend, lower modules to higher modules, especially do not miss project files like `package.json` and so on.
-    
-        An example of file structure:
-        ```json: files.json
-        [
-          "backend/package.json",
-          "backend/server.js",
-          ...
-          "frontend/package.json"
-        ]
-        ```
-        
-        The format is different(```json: files.json is fixed), then `files.json` will be saved on the file system automatically.
-    
-    4. Additional instructions:
-      * Do not use conflict ports for different projects, use obscure port for each of them
-      * Present all information in a concise yet comprehensive manner
-      * Focus on the architect job, do not write code yourself, the coding task will come later
-
-    Now Begin:
-
-callbacks:
-  - callbacks/artifact_callback
-
-max_chat_round: 1
-
-output_dir: output
-
-help: |
