--- projects/code_scratch_test2/workflow.yaml
@@ -1,16 +1,0 @@
-architecture:
-  next:
-    - coding
-  agent_config: architecture.yaml
-  agent:
-    name: LLMAgent
-    kwargs:
-      tag: Architecture
-
-
-coding:
-  agent_config: coding.yaml
-  agent:
-    name: LLMAgent
-    kwargs:
-      tag: Coding
